refactor(crypto): move resolvePriceDecimals to shared/lib + e2e timeout consts (PR #623 review R2)
- Move pure resolvePriceDecimals from widgets/chart/utils to shared/lib/priceFormat (co-locate with dynamicDecimals)
- Update StockChart import and remove old widget-local utility
- Extract E2E timeout magic numbers to named consts in crypto-symbol.spec.ts
- Consolidate duplicate vitest imports across 4 guard test files
