+++ src/app/[symbol]/fundamental/__tests__/page.guard.test.ts
@@ -101,8 +101,14 @@ vi.mock('react-error-boundary', () => ({
     ErrorBoundary: ({ children }: { children: React.ReactNode }) => children,
 }));
 
-import { describe, expect, it, beforeEach, vi } from 'vitest';
-import type { MockedFunction } from 'vitest';
+import {
+    describe,
+    expect,
+    it,
+    beforeEach,
+    vi,
+    type MockedFunction,
+} from 'vitest';
 import { NOINDEX_SYMBOL_METADATA } from '@/shared/lib/seo';
 import { isTabAllowedForSymbol } from '@/entities/ticker/api';
 import { getAssetInfoResilient } from '@/entities/ticker';
