+++ src/shared/lib/priceFormat.ts
@@ -1,4 +1,8 @@
-import type { PriceFormatConfig } from '@/shared/config/marketProfile';
+import { getDescriptor } from '@/shared/config/marketProfile';
+import type {
+    MarketProfileId,
+    PriceFormatConfig,
+} from '@/shared/config/marketProfile';
 
 type PriceSign = '+' | '';
 type PriceArrow = '▲' | '▼';
@@ -83,4 +87,30 @@ export function formatPriceChange(percent: number): PriceChangeDisplay {
         arrow: isUp ? '▲' : '▼',
         arrowLabel: isUp ? '상승' : '하락',
     };
+}
+
+/**
+ * Resolve the number of price decimals for a market profile.
+ *
+ * fixed/integer descriptors are static; dynamic (crypto) derives significant
+ * digits from the latest close magnitude so sub-cent tokens aren't flattened.
+ *
+ * Placed in priceFormat (rather than widgets/) because it is a pure function
+ * that depends only on shared/config/marketProfile and dynamicDecimals — no
+ * React or widget graph dependency. Co-locating with dynamicDecimals keeps
+ * related price-formatting utilities together and allows shared/ consumers to
+ * import it without a cross-layer violation.
+ *
+ * @param marketProfile - The market profile id (e.g. 'us-equity', 'crypto').
+ * @param lastClose - The most recent close price; used only for 'dynamic-by-magnitude'
+ *   precision. Defaults to 1 (returns 2 decimals) when undefined.
+ */
+export function resolvePriceDecimals(
+    marketProfile: MarketProfileId,
+    lastClose: number | undefined
+): number {
+    const precision = getDescriptor(marketProfile).priceFormat.precision;
+    if (precision.kind === 'fixed') return precision.digits;
+    if (precision.kind === 'integer') return 0;
+    return dynamicDecimals(lastClose ?? 1);
 }
