+++ src/widgets/chart/StockChart.tsx
@@ -67,7 +67,7 @@ import {
     MA_DEFAULT_PERIODS,
 } from '@y0ngha/siglens-core';
 import type { MarketProfileId } from '@/shared/config/marketProfile';
-import { resolvePriceDecimals } from './utils/resolvePriceDecimals';
+import { resolvePriceDecimals } from '@/shared/lib/priceFormat';
 import { IndicatorSettingsModal } from './ui/IndicatorSettingsModal';
 import {
     INDICATOR_META,
